stop tracking the dumb ./build/

+++ src/Cinematic.cpp
@@ -16,7 +16,7 @@ void Cinematic::draw(SDL_Renderer* renderer) {
 	Presenter::drawObject(background);
 	for (int i = cropRects.size() - 1; i >= 0; i--) {
 		double swayingAngle = 5 * sin(0.07 * ticks + i);
-		SDL_Point swayingPoint = { cropRects[i].w * 0.32, cropRects[i].h * 0.97 };
+		SDL_Point swayingPoint = { (int)(cropRects[i].w * 0.32), (int)(cropRects[i].h * 0.97) };
 		SDL_RenderCopyEx(renderer, cropTexture, NULL, &cropRects[i], swayingAngle, &swayingPoint, SDL_FLIP_NONE);
 	}
 	ticks++;
